did small changes

+++ adjoint/my_bfgs/LmemoryHessian.py
@@ -139,10 +139,10 @@ class LimMemoryHessian(InvertedHessian):
         rhok, yk, sk = self[k]
         
         #print rhok,yk,sk
-        t = x - rhok * x.dot(sk) * yk
+        t = x - float(rhok * x.dot(sk)) * yk
         t = self.matvec(t,k-1)
-        t = t - rhok * yk.dot(t) * sk
-        t = t + rhok * x.dot(sk) * sk
+        t = t - float(rhok * yk.dot(t)) * sk
+        t = t + float(rhok * x.dot(sk)) * sk
         return t
